Initial commit moving utility functions for parsing and upload to a utilities.py file. Moved functionality for login, signup, upload also to this file to streamline server.py.
